requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0.22
pydantic>=2.4.2
pydantic-settings>=2.0.3
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
pytest>=7.4.3
httpx>=0.25.0
